First steps towards a real bootloader
Removed all the game code and set up a FAT file system with the
bootloader.asm.

Next up would be implementing reading the FAT system (C(++) or MASM) for
a 1MB stage2 bootloader file which would be loaded at 0x0 and then jump
to it.

This would nullify the current size contraints and could be implemented
in any language. I haven't decided yet between rust (want to use it for
the kernel in any case) and C(++) (to be consistent within the
bootloader project).